Refactor: Move `scoped_process` to _stbt.utils
So that I can use it in other tests in a future commit.
